Add morpheus_cloud_folder data-source and add folderId to instance
In some circumstances the VMware Folder ID needs to be provided when
creating an instance.  We add "folder_id" as an optional parameter to
the configuration for morpheus_vsphere_instance.  We also add a
data-source morpheus_cloud_folder to get the ID of a cloud folder.
This data-source can be used to get a VMware Folder ID.

We've regenerated documentation for the provider, which includes docs
for the new data-source.