Return on error in morpheus_vsphere_instance
A couple of customers have reported stack-traces from
morpheus_vsphere_instance.  The reason for these stack-traces is that
GETs of specific resources return 403s, but instead of returning with an
error the code continues to execute and will eventually try to access a
field that is nil.  To fix this we make two sets of changes:
- return if there is a GET error
- check the return code of various casts, and if not ok then return with
  an error