Add instance_type_code to morpheus_vsphere_instance
Some customers want to use terraform to provision instances using
sub-tenant users who may not have sufficient privilege to GET
information on instance-types in the master tenant.  To accommodate
this use-case we've added "instance_type_code" to the Schema of
morpheus_vsphere_instance.  This will bypass the logic in
morpheus_vsphere_instance to GET instance-type information and retrieve
the code from it.  We're using "ExactlyOneOf" in the Schema to make sure
that only one of "instance_type_code" or "instance_type_id" are
specified when creating a morpheus_vsphere_instance, and that one of
them is always specified.