Remove GET of instance-layout ID from morpheus_vsphere_instance
The request body of the POST to create an instance takes the
instance-layout ID and no other instance-layout information.  In this PR
we remove the GET of instance-layout ID performed in advance of the POST
in morpheus_vsphere_instance.

We also add instance_type_code to the state-file, an oversight from a
previous PR.