Add connection_info to morpheus_vsphere_instance
Some customers want to get the IP address of a created instance for use
as inventory for Ansible (for example).  The API spec includes a
"connection_info" list which from testing is populated with the IP
address and port to be used in connecting to a network interface.  In
this PR we add "connection_info" to the Schema for
morpheus_vsphere_instance and populate it with values from the reponse
to the POST request.

We also add the interface information returned from the POST to the
state-file.