Only set one of instance_type_code or instance_type_id in state-file
We introduced a bug in v0.13.1 whereby we only allowed the user to set
one of instance_type_code or instance_type_id in the config for
morpheus_vsphere_instance but we stored the value of both in the
state-file.  To fix this in this PR we:
- add logic to morpheus_vsphere_instance to only set the value of
  instance_type_code in the state-file if there is a change
- add logic to do the same with instance_type_id
The logic is added to the read method.

This code will repair the bug in existing state-files and ensure that
the bug will not reoccur.  This means that on the first run of apply
with where a morpheus_vsphere_instance is present in the state-file an
update will be performed which won't change the instance but which will
fix the state-file (on the read).  Subsequent runs of apply will do
nothing.

We already ensure that a recreate will be triggered if either
instance_type_code or instance_type_id are actually changed by the user
through the use of CusomizeDiff functions.