Fix a number of issues with mvm_instance
We've seen issues with morpheus_mvm_instance where volumes can be
returned with some null values.  This results in a stack-trace. Also
ids can be returned as strings. In this PR we check that volume values
are non-nil before type assertions and we will also convert a
datastoreId string to int.

We also check that the volumes set in the config are in the returned
list from the API.

We also check that the evars set in the config are in the returned list
from the API.

We don't set volumes or evars to anything returned by the API. Except
that we set the volume uuid to the value returned by the API.

We've added an error message if the resourcePoolId loop can't find an id
that matches the name and has a providerType of mvm.